refactor(verb): Rename ansi to verb and move ANSI color functions
The `ansi` package has been renamed to `verb` and its ANSI color
functions have been moved into this package. This change simplifies the
import structure and consolidates color-related functionality within the
`verb` package. The usage of these functions has been updated
accordingly.
